requirements.txt
telethon
googletrans
pillow
pytesseract